+++ src/engine.cpp
@@ -18,8 +18,16 @@
 #include "stackFrame.h"
 
 
+Error Engine::check(Arguments& args) {
+    return Error::OK;
+}
+
+bool Engine::requireNativeTrace() {
+    return true;
+}
+
 int Engine::getNativeTrace(void* ucontext, int tid, const void** callchain, int max_depth,
-                           const void* jit_min_address, const void* jit_max_address) {
+                           CodeCache* java_methods, CodeCache* runtime_stubs) {
     StackFrame frame(ucontext);
     const void* pc = (const void*)frame.pc();
     uintptr_t fp = frame.fp();
@@ -32,7 +40,7 @@ int Engine::getNativeTrace(void* ucontext, int tid, const void** callchain, int 
     while (depth < max_depth && pc >= valid_pc) {
         callchain[depth++] = pc;
 
-        if (pc >= jit_min_address && pc < jit_max_address) {
+        if (java_methods->contains(pc) || runtime_stubs->contains(pc)) {
             break;
         }